add locations map to dashboard
