+++ src/views/partials/common/menu.hbs
@@ -11,6 +11,9 @@
         <a class="navbar-item" href="/dashboard">
           Dashboard
         </a>
+        <a class="navbar-item" href="/my-movies">
+          My Movies
+        </a>
         {{#if user.isAdmin}}
         <a class="navbar-item" href="/admin">
           Admin
